Rename solution and projects / adjust namespaces and assembly names
