+++ Samples/1_NET_Core/NetCoreConsoleSample.Domain/Commands/CreateCustomerFromConsoleCommand.cs
@@ -1,6 +1,6 @@
 ﻿using BE.CQRS.Domain.Commands;
 
-namespace Domain.Commands
+namespace NetCoreConsoleSample.Domain.Commands
 {
     public sealed class CreateCustomerFromConsoleCommand : ICommand // All Commands must be dervied from ICommand
     {
