+++ Samples/1_NET_Core/NetCoreConsoleSample.Domain/Events/CustomerCreatedFromConsoleEvent.cs
@@ -1,6 +1,6 @@
 ﻿using BE.CQRS.Domain.Events;
 
-namespace Domain.Events
+namespace NetCoreConsoleSample.Domain.Events
 {
     public sealed class CustomerCreatedFromConsoleEvent : EventBase
     {
