+++ Samples/1_NET_Core/NetCoreConsoleSample.Domain/NetCoreConsoleSample.Domain.csproj
@@ -2,6 +2,9 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
+    <AssemblyName>NetCoreConsoleSample.Domain</AssemblyName>
+    <RootNamespace>NetCoreConsoleSample.Domain</RootNamespace>
+    <LangVersion>7.3</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
